EditRecipeScreen finished, added EditRecipeViewModel
